[add] tangential obstacle

+++ src/resources/DimensionBundle.java
@@ -21,6 +21,6 @@ public class DimensionBundle extends ListResourceBundle {
             {"LargeRingObstacle_radius", new Double(100.0)},
             {"Obstacle_spacing", new Double(350.0)},
             {"Windmill_radius", new Double(107.5)},
-            {"Num_of_obstacles", new Integer(2)} // update manually when adding new obstacle
+            {"Num_of_obstacles", new Integer(3)} // update manually when adding new obstacle
     };
 }
